// src/app/shared/enum/entity.enum.ts
export const enum EntityEnum {
  loading = 'loading',
  property = 'property',
  promotion = 'promotion',
  contact = 'contact',
  lead = 'lead',
  task = 'task',
  location = 'location',
  locationPath = 'locationPath',
  locationCountry = 'locationCountry',
  locationCanton = 'locationCanton',
  locationDistrict = 'locationDistrict',
  locationZone = 'locationZone',
  locationCity = 'locationCity',
  locationQuarter = 'locationQuarter',
  matching = 'matching',
  matchingGroup = 'matchingGroup',
  reporting = 'reporting',
  sector = 'sector',
  report = 'report',
  broker = 'broker',
  intermediate = 'intermediate',
  reportContact = 'reportContact',
  agencyPreference = 'agencyPreference',
  marketingExpense = 'marketingExpense',
  emailSent = 'emailSent',
  portal = 'portal',
  device = 'device',
  customAttribute = 'customAttribute',
  dns = 'dns',
  emailTemplate = 'emailTemplate',
  contract = 'contract',
  agency = 'agency',
  user = 'user',
  document = 'document',
  offer = 'offer',
  emailing = 'emailing',
  recipient = 'recipient',
  restriction = 'restriction',
  suggestion = 'suggestion',
  suggestionTag = 'suggestionTag',
  help = 'help',
  website = 'website',
  websiteArticle = 'websiteArticle',
  agencyBackup = 'agencyBackup',
  agencyProfile = 'agencyProfile',
  agenda = 'agenda',
  cmsArticle = 'cmsarticle',
  mls = 'mls',
  production = 'production',
  account = 'account',
  cms = 'cms',
  process = 'process',
}